crudを分解
