+++ app/repositories/user.py
@@ -1,0 +1,22 @@
+import hashlib
+
+from sqlalchemy.orm import Session
+
+from app.models.user import User
+from app.schemas import schemas
+
+
+def get_users(db: Session, skip: int = 0, limit: int = 100):
+    return db.query(User).offset(skip).limit(limit).all()
+
+
+def create_user(db: Session, user: schemas.UserCreate):
+    db_user = User(
+        name=user.name,
+        email=user.email,
+        password=hashlib.sha256(user.password.encode()).hexdigest()
+    )
+    db.add(db_user)
+    db.commit()
+    db.refresh(db_user)
+    return db_user
